+++ src/presentation/legallead.desktop.core/entities/UserBo.cs
@@ -6,7 +6,7 @@ namespace legallead.desktop.entities
     {
         private AccessTokenBo? token;
 
-        public bool IsAuthenicated => Token != null && Token.Expires.HasValue;
+        public virtual bool IsAuthenicated => Token != null && Token.Expires.HasValue;
         public string UserName { get; set; } = string.Empty;
         public ApiContext[]? Applications { get; set; }
         public bool IsInitialized => Applications != null && Applications.Length > 0;
