+++ src/presentation/legallead.desktop.core/implementations/UserPermissionsMapper.cs
@@ -1,7 +1,6 @@
 ﻿using HtmlAgilityPack;
 using legallead.desktop.entities;
 using legallead.desktop.interfaces;
-using System.Runtime.Intrinsics.X86;
 
 namespace legallead.desktop.implementations
 {
