+++ src/presentation/legallead.desktop.core/interfaces/IUserPermissionsMapper.cs
@@ -1,9 +1,4 @@
 ﻿using legallead.desktop.entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace legallead.desktop.interfaces
 {
