+++ src/presentation/legallead.desktop.tests/implementations/CopyrightBuilderTests.cs
@@ -1,9 +1,4 @@
 ﻿using legallead.desktop.implementations;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace legallead.desktop.tests.implementations
 {
